Moved fitting and datastorage to qudi.util namespace
