+++ qudi/core/gui/qtwidgets/fitting.py
@@ -23,7 +23,7 @@ top-level directory of this distribution and at <https://github.com/Ulm-IQO/qudi
 import os
 import weakref
 from PySide2 import QtCore, QtWidgets, QtGui
-from qudi.core.datafitting import FitContainer, FitConfigurationsModel, FitConfiguration
+from qudi.util.datafitting import FitContainer, FitConfigurationsModel, FitConfiguration
 from qudi.core.paths import get_artwork_dir
 from qudi.core.gui.qtwidgets.scientific_spinbox import ScienDSpinBox
