+++ qudi/util/fit_models/sine.py
