+++ setup.py
@@ -86,8 +86,8 @@ setup(name='qudi',
                 'qudi.core.jupyterkernel',
                 'qudi.core.jupyterkernel.mpl',
                 'qudi.util',
+                'qudi.util.fit_models',
                 'qudi.tools',
-                'qudi.tools.fit_models',
                 'qudi.tools.config_editor'
                 ],
       package_data={'': ['LICENSE.txt', 'COPYRIGHT.txt', 'docs/*'],
