+++ tests/fit_methods/test_gaussian.py
@@ -23,7 +23,7 @@ top-level directory of this distribution and at <https://github.com/Ulm-IQO/qudi
 import unittest
 import numpy as np
 
-from qudi.core.fit_models.gaussian import Gaussian
+from qudi.util.fit_models.gaussian import Gaussian
 
 
 class TestGaussianMethods(unittest.TestCase):
