+++ tests/fit_methods/test_lorentzian.py
@@ -23,7 +23,7 @@ top-level directory of this distribution and at <https://github.com/Ulm-IQO/qudi
 import unittest
 import numpy as np
 
-from qudi.core.fit_models.lorentzian import Lorentzian
+from qudi.util.fit_models.lorentzian import Lorentzian
 
 
 class TestLorentzianMethods(unittest.TestCase):
